refactor profile view
