+++ components/ui/profile/ProfileRefreshControl.tsx
@@ -1,0 +1,18 @@
+import React from "react";
+import { RefreshControl } from "react-native";
+import { UseProfile } from "../../hooks/profile/useProfile";
+
+interface IProps {
+  profile: UseProfile;
+}
+
+export function ProfileRefreshControl({ profile }: IProps) {
+  return (
+    <RefreshControl
+      refreshing={profile.refreshing}
+      onRefresh={() => {
+        profile.doLoad(true).then();
+      }}
+    />
+  );
+}
